# api-contract.yml
openapi: 3.0.3
info:
  title: Sistema Stock - API Contract
  description: |
    Contrato de API para el sistema de gestión de stock.
    
    ## Arquitectura
    - **Backoffice**: Gestión interna con service_role key
    - **B2B**: Tienda para clientes con auth JWT
    - **OCR**: Microservicio externo en Hetzner
    
    ## Autenticación
    - **Backoffice**: Service role key en headers
    - **B2B**: JWT token de usuario autenticado
    - **OCR**: HTTP Basic Authentication
    
    ## Base URLs
    - **Vercel Apps**: `https://your-app.vercel.app`
    - **OCR Service**: `https://your-ocr.hetzner.com`
  version: 1.0.0
  contact:
    name: Sistema Stock Team
    email: dev@sistema-stock.com
servers:
  - url: https://sistema-stock-b2b.vercel.app/api
    description: B2B Production API
  - url: https://sistema-stock-backoffice.vercel.app/api
    description: Backoffice Production API
  - url: https://ocr-service.hetzner.com
    description: OCR Service (Hetzner)
  - url: http://localhost:3000/api
    description: Development Local
security:
  - BearerAuth: []
  - ServiceRoleAuth: []
  - BasicAuth: []

# =========================================================
# MODELOS DE DATOS
# =========================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token para usuarios B2B autenticados
    
    ServiceRoleAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "Bearer {SUPABASE_SERVICE_ROLE_KEY} para Backoffice"
    
    BasicAuth:
      type: http
      scheme: basic
      description: HTTP Basic para servicio OCR
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nombre:
          type: string
          example: "Tomate Cherry"
        descripcion:
          type: string
        unidad:
          type: string
          enum: [kg, g, l, ml, ud, pz, caja, bolsa]
          default: kg
        stock_actual:
          type: number
          format: decimal
          minimum: 0
          example: 25.5
        stock_minimo:
          type: number
          format: decimal
          minimum: 0
        stock_maximo:
          type: number
          format: decimal
        categoria:
          type: string
          example: "verduras"
        proveedor_principal:
          type: string
          format: uuid
        precio_promedio:
          type: number
          format: decimal
          minimum: 0
          example: 3.80
        brand:
          type: string
        image_url:
          type: string
          format: uri
        is_active:
          type: boolean
          default: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProductWithDiscount:
      allOf:
        - $ref: '#/components/schemas/Product'
        - type: object
          properties:
            discount_percentage:
              type: number
              format: decimal
              minimum: 0
              maximum: 100
              example: 10.5
            final_price:
              type: number
              format: decimal
              minimum: 0
              example: 3.42
...
    Entry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tipo:
          type: string
          enum: [ocr, manual]
        estado:
          type: string
          enum: [draft, processing, completed, error, validated]
        supplier_id:
          type: string
          format: uuid
        proveedor_text:
          type: string
        fecha_factura:
          type: string
          format: date
        numero_factura:
          type: string
        documento_url:
          type: string
          format: uri
        archivo_nombre:
          type: string
        productos:
          type: array
          items:
            type: object
        ocr_confidence:
          type: number
          format: decimal
          minimum: 0
          maximum: 1
        total_items:
          type: integer
          minimum: 0
        total_amount:
          type: number
          format: decimal
          minimum: 0
        validated_by:
          type: string
          format: uuid
        validated_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    OCRResponse:
      type: object
      properties:
        success:
          type: boolean
        processing_id:
          type: string
          format: uuid
        proveedor:
          type: object
          properties:
            nombre:
              type: string
            direccion:
              type: string
            telefono:
              type: string
            email:
              type: string
              format: email
            cif:
              type: string
            confianza:
              type: number
              minimum: 0
              maximum: 1
        documento:
          type: object
          properties:
            tipo:
              type: string
              enum: [factura, albarán]
            numero:
              type: string
            fecha:
              type: string
              format: date
            total:
              type: number
              minimum: 0
        productos:
          type: array
          items:
            type: object
            properties:
              nombre:
                type: string
              cantidad:
                type: number
                minimum: 0.001
              unidad:
                type: string
              precio:
                type: number
                minimum: 0
              precio_total:
                type: number
                minimum: 0
              caducidad:
                type: string
                format: date
                nullable: true
              lote:
                type: string
                nullable: true
              confianza:
                type: number
                minimum: 0
                maximum: 1
        texto_completo:
          type: string
        confianza_general:
          type: number
          minimum: 0
          maximum: 1
        tiempo_procesamiento:
          type: number
          minimum: 0
        error:
          type: string
          nullable: true

    # =========================================================
    # RESPUESTAS COMUNES
    # =========================================================
...
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items: {}
        pagination:
          type: object
          properties:
            page:
              type: integer
              minimum: 1
            limit:
              type: integer
              minimum: 1
              maximum: 100
            total:
              type: integer
              minimum: 0
            total_pages:
              type: integer
              minimum: 0

# =========================================================
# ENDPOINTS API
# =========================================================
paths:
...
  /stock:
    get:
      tags: [B2B - Stock]
      summary: Obtener catálogo de productos con descuentos
      description: |
        Lista productos activos con stock disponible, aplicando descuentos del cliente autenticado.
        Incluye precio final calculado con descuentos por producto/categoría.
      security:
        - BearerAuth: []
      parameters:
        - name: category
          in: query
          description: Filtrar por categoría
          schema:
            type: string
        - name: search
          in: query
          description: Búsqueda por nombre
          schema:
            type: string
        - name: in_stock
          in: query
          description: Solo productos con stock > 0
          schema:
            type: boolean
            default: true
        - name: page
          in: query
          description: Número de página
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Items por página
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Lista de productos con descuentos aplicados
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ProductWithDiscount'
        '401':
          description: Usuario no autenticado
        '403':
          description: Cliente no aprobado

  # =========================================================
  # B2B - PEDIDOS
  # =========================================================
...
  /entries:
    get:
      tags: [Backoffice - Entries]
      summary: Listar entradas de mercancía
      security:
        - ServiceRoleAuth: []
      parameters:
        - name: estado
          in: query
          schema:
            type: string
            enum: [draft, processing, completed, error, validated]
        - name: tipo
          in: query
          schema:
            type: string
            enum: [ocr, manual]
        - name: supplier_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista de entradas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Entry'

    post:
      tags: [Backoffice - Entries]
      summary: Crear entrada manual
      description: Crear entrada manual de mercancía sin OCR
      security:
        - ServiceRoleAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tipo, productos]
              properties:
                tipo:
                  type: string
                  enum: [manual]
                supplier_id:
                  type: string
                  format: uuid
                fecha_factura:
                  type: string
                  format: date
                numero_factura:
                  type: string
                productos:
                  type: array
                  items:
                    type: object
                    required: [nombre, cantidad]
                    properties:
                      nombre:
                        type: string
                      cantidad:
                        type: number
                        minimum: 0.001
                      precio:
                        type: number
                        minimum: 0
                      caducidad:
                        type: string
                        format: date
                      lote:
                        type: string
      responses:
        '201':
          description: Entrada creada exitosamente
...
  /ocr/callback:
    post:
      tags: [Backoffice - OCR]
      summary: Callback desde servicio OCR
      description: |
        Endpoint interno para recibir resultados del servicio OCR.
        Actualiza la entrada con datos extraídos.
      security:
        - ServiceRoleAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OCRResponse'
      responses:
        '200':
          description: Callback procesado exitosamente

  # =========================================================
  # BACKOFFICE - ALERTAS
  # =========================================================
...